Split test class by testing methods
